+++ database/migrations/2024_09_20_164359_create_sist_bitacora_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id('idbitacora');
             $table->string('tabla_modificada', 150);
             $table->integer('objeto_modificado');
-            $table->unsignedBigInteger('idusuario'); // Asegúrate de que el tipo de datos coincida con la tabla de usuarios
+            $table->integer('idusuario'); // Asegúrate de que el tipo de datos coincida con la tabla de usuarios
             $table->dateTime('fecha');
             $table->string('accion', 45);
         });
